feat: refactor code to separate img format conversion
The macOS flow created raw images while the Linux flow created a VMDK.
The code and comments all said VMDK. Updated the code to create the raw
disk image in a temporary path which gets cleaned up automatically for
us while the image conversion is done after the fact. Set the default
format to raw for now as well.