fix: bad path generation for netinit
The installed path is esxi_netinit which is where the files were but the
directory entry was hardcoded to netinit which was wrong. GNU tar didn't
have a problem with this but ESXi 8 did. Refactored the code to walk the
tarball in one go instead of two separate loops. To avoid problems like
this in the future, introduce a helper class to ensure our directories
are created and the files are properly sorted into each directory.