fix: refactor network config to set MAC and MTU and DHCP
The network configuration that this script aims to bring up is the
management network so that someone can log into the hypervisor and then
finish configuring it. This makes adjustments to how the IP interfaces
are assigned to include their MAC address and the MTU for the network to
bring up the networks. Updates the configuration of the interfaces to
support DHCP as well as static IPv4 addresses.